% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexForecast_model.R, R/xflex4cast_model.R
\name{define_xflex4cast}
\alias{define_xflex4cast}
\title{Define xflex4cast Model Parameters}
\usage{
define_xflex4cast(
  mqgam_formula,
  mqgam_quantiles = c(0.05, 0.25, 0.5, 0.75, 0.95),
  fit_tail = TRUE,
  scale_formula = ~1,
  quantile_threshold = 0.95,
  tail_distribution = "DGP",
  maximum_quantile_level = 0.999
)

define_xflex4cast(
  mqgam_formula,
  mqgam_quantiles = c(0.05, 0.25, 0.5, 0.75, 0.95),
  fit_tail = TRUE,
  scale_formula = ~1,
  quantile_threshold = 0.95,
  tail_distribution = "DGP",
  maximum_quantile_level = 0.999
)
}
\arguments{
\item{mqgam_formula}{A formula for the MQGAM model.}

\item{mqgam_quantiles}{Quantile levels to be fitted. Default: c(0.05, 0.25, 0.5, 0.75, 0.95). If \code{fit_tail=TRUE} values
above the quantile_threshold will be modelled by the DGP, while below to it the \code{mqgam}.}

\item{fit_tail}{Logical, whether to fit a tail model for high quantiles. Default: TRUE.}

\item{scale_formula}{Formula used in tail model (GJRM). Default: ~ 1.}

\item{quantile_threshold}{Threshold quantile for tail modeling. Default: 0.95.}

\item{tail_distribution}{String indicating the family to be used in tail modeling. Default: "DGP".}

\item{maximum_quantile_level}{Maximum quantile level to be used. Default: 0.999.}
}
\value{
A list with class \code{xflex4cast}.

A list with class \code{xflex4cast}.
}
\description{
Internal function to define and initialize parameters for an xflex4cast model.

Internal function to define and initialize parameters for an xflex4cast model.
}
\keyword{internal}
