% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xflex4cast_model.R
\name{fit.xflex4cast}
\alias{fit.xflex4cast}
\title{Fit xflex4cast Model}
\usage{
\method{fit}{xflex4cast}(model, data, index = NULL, fit_tail_only = FALSE)
}
\arguments{
\item{model}{A \code{xflex4cast} object.}

\item{data}{A \code{data.frame} or \code{data.table} with the training data.}

\item{index}{Column name for unique identifier used in cross-validation. If \code{NULL}, uses row numbers.}

\item{fit_tail_only}{boolean. If \code{TRUE} it will use the currently fitted \code{mqgam_fit} model object from the
\code{xflex4cast} object and fit only the tail.}
}
\value{
A fitted \code{xflex4cast} object.
}
\description{
Internal method to fit a \code{xflex4cast} model using MQGAM and optionally fit a tail model.
}
