% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_computations.R
\name{get_qgp_using_parameters}
\alias{get_qgp_using_parameters}
\title{Function to get the quantiles of a Generalised Pareto distribution
as a function of the shape and scale parameters.}
\usage{
get_qgp_using_parameters(p, p4r_tail, sigma_new = NULL, shape = NULL)
}
\arguments{
\item{p}{probability level associated with the quantile}

\item{sigma_new}{a numeric value or vector for the scale parameter}

\item{shape}{a constant value for the shape parameter}

\item{p4r_mqgam}{A \code{p4r_tail} object that has been fit.}
}
\value{
a quantile associated with the \code{p} probability levelfor the
\code{p4r_mqgam} adjusted model as function of shape and scale parameter.
}
\description{
Function to get the quantiles of a Generalised Pareto distribution
as a function of the shape and scale parameters.
}
