% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_degpd.R
\name{load_degpd}
\alias{load_degpd}
\title{Load DEGPD and ZIDEGPD R and C++ Functions}
\usage{
load_degpd()
}
\value{
Invisibly returns \code{NULL}. Used for its side effects.
}
\description{
Dynamically loads all R and C++ source files required for the DEGPD and ZIDEGPD models.
}
\details{
This function sources all R files from \verb{inst/degpd-and-zidegpd/R/} and attempts to
compile and load all C++ files via \code{Rcpp::sourceCpp()} from \verb{inst/degpd-and-zidegpd/src/}.
It is designed to be used during package development or advanced workflows where
direct loading of source files is needed.

The function assumes the existence of two subdirectories within the \verb{inst/degpd-and-zidegpd/} path:
\itemize{
\item \verb{R/}: containing R source files to be loaded via \code{source()}.
\item \verb{src/}: containing C++ source files to be compiled and loaded via \code{Rcpp::sourceCpp()}.
}

If any C++ file fails to compile, the function will suppress the error using \code{try()}, and
continue executing. This may result in incomplete loading of required functionality.

This function is intended for internal use, particularly useful in development environments. For more details
check the original repository: https://github.com/touqeerahmadunipd/degpd-and-zidegpd; and mail the original
author at touqeer.ahmad8960@gmail.com.
}
\examples{
\dontrun{
# Load all DEGPD and ZIDEGPD functions (R and C++)
load_degpd()
}
}
\references{
Ahmad, T., Gaetan, C., & Naveau, P. (2024). An extended generalized Pareto regression model for count data. \emph{Statistical Modelling}, 0(0). https://doi.org/10.1177/1471082X241266729
}
\seealso{
\code{\link[Rcpp:sourceCpp]{Rcpp::sourceCpp()}}, \code{\link[base:source]{base::source()}}, \code{\link[=list.files]{list.files()}}
}
