% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{mixture_dgamma_constant}
\alias{mixture_dgamma_constant}
\title{Generate Data from a Two-Component Discrete Mixture with Fixed Cut-off}
\usage{
mixture_dgamma_constant(
  n = 100,
  phi = 0.05,
  xi = 0.3,
  scale_dgp = 15,
  shape_dgamma = 5,
  scale_dgamma = 4,
  probs = c(0.001, 0.01, 0.05, 0.25, 0.5, 0.75, 0.8, 0.9, 0.95, 0.99, 0.999, 0.9999,
    0.99999),
  index = NULL
)
}
\arguments{
\item{n}{Integer. Number of observations to simulate.}

\item{phi}{Numeric. Tail probability beyond which the DGP distribution is used.}

\item{xi}{Numeric. Shape parameter for the DGP distribution.}

\item{scale_dgp}{Numeric. Scale parameter for the DGP distribution.}

\item{shape_dgamma}{Numeric. Shape parameter for the discrete gamma distribution.}

\item{scale_dgamma}{Numeric. Fixed scale parameter for the gamma distribution.}

\item{probs}{Numeric vector. Quantile levels for which to compute the distribution quantiles.}

\item{index}{Optional. If not NULL, an index variable will be added to the output.}
}
\value{
A list with:
\describe{
\item{data}{A \code{data.table} with covariates and generated outcome \code{y}.}
\item{pmf_dggp}{Named vector containing the PMF of the full mixture.}
\item{true_quantiles}{Vector of quantiles corresponding to \code{probs}.}
}
}
\description{
Simulates data from a discrete mixture where the lower tail is from a discrete gamma distribution and the upper tail is filled by a discrete Generalized Pareto (DGP) beyond a fixed quantile threshold.
}
\details{
This function defines a fixed upper-tail threshold using the gamma quantile and then reweights the DGP to fill the tail such that the mixture retains a valid PMF.
}
\examples{
out <- mixture_dgamma_constant()
head(out$data)

}
