% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_computations.R
\name{pdgp}
\alias{pdgp}
\title{Cumulative distribution function for the Discrete Generalised Pareto distribution}
\usage{
pdgp(q, shape, scale)
}
\arguments{
\item{q}{quantiles}

\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\description{
Cumulative distribution function for the Discrete Generalised Pareto distribution
}
\details{
Internal use.
}
