% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_computations.R
\name{pgp}
\alias{pgp}
\title{Cumulative distribution function for the continuous Generalised Pareto distribution}
\usage{
pgp(q, shape, scale)
}
\arguments{
\item{q}{quantiles}

\item{shape}{shape parameter}

\item{scale}{scale parameter}
}
\description{
Cumulative distribution function for the continuous Generalised Pareto distribution
}
\details{
Internal use.
}
