% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{pinball}
\alias{pinball}
\title{Pinball Loss}
\usage{
pinball(predictions, data, valid_time_index = NULL, indexes = NULL)
}
\arguments{
\item{predictions}{probabilistic fault forecasts, quantiles in long
format \code{data.table}.}

\item{data}{actual fault data with columns \code{indexes} and column \code{actual}
containing outcome corresponding to predictions.}

\item{valid_time_index}{common column name containing the index on which
to merge \code{predictions} and \code{data}. If \code{NULL}, a merge will be attempted on
if there a single common column name exists.}

\item{indexes}{column names in \code{predictions} used to group average pinball
scores by.}
}
\value{
a \code{data.table} of pinball scores with the same indexes as \code{predictions}
(dtm(s) and quantiles).
}
\description{
This function takes as input a data.table of predictions
(with columns \code{quantile} and \code{value}) and a \code{data.table} of
actual values (with column \code{actual}). Both should contain an index column
to be used for merging, which may be specified using \code{valid_time_index}. It computes
the average pinball loss grouped by quantiles and \code{indexes} and returns the
results as a \code{data.table}.
}
