% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_computations.R
\name{qgp}
\alias{qgp}
\title{Calculate the quantile for a Generalised Pareto Distribution (Continuous or
Discrete) fitted through a GAMLSSS model.}
\usage{
qgp(p, newdata, p4r_tail, sigma_new = NULL, shape = NULL)
}
\arguments{
\item{p}{probability level associated with the quantile}

\item{newdata}{Input data to be used to generate new predictions.}

\item{p4r_tail}{A \code{p4r_tail} object that has been fit.}
}
\value{
a quantile associated with the \code{p} probability level for exceedances
from a \code{p4r_tail} fit.
}
\description{
Calculate the quantile for a Generalised Pareto Distribution (Continuous or
Discrete) fitted through a GAMLSSS model.
}
