% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quantile_matrix}
\alias{quantile_matrix}
\title{Convert prediction table into day-by-quantile matrix}
\usage{
quantile_matrix(dt)
}
\arguments{
\item{dt}{A data.table with columns: quantile, valid_time, value}
}
\value{
A numeric matrix with rows = valid_time and columns = quantile
}
\description{
Convert prediction table into day-by-quantile matrix
}
