% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{reliability}
\alias{reliability}
\title{Reliability/Calibration}
\usage{
reliability(
  predictions,
  data,
  valid_time_index = NULL,
  indexes = NULL,
  exclude_zero_pred = T
)
}
\arguments{
\item{predictions}{probabilistic fault forecasts, quantiles in long format
\code{data.table}.}

\item{data}{actual fault data with columns \code{indexes} and column \code{actual}
containing outcome corresponding to predictions.}

\item{valid_time_index}{common column name containing the index on which
to merge \code{predictions} and \code{data}.}

\item{indexes}{column names in \code{predictions} used to group result by,
defaults to \code{NULL} (no grouping)}

\item{exclude_zero_pred}{exclude predicted quantiles that equal zero. Recommended
of the target variable is zero-inflated. NB: other treatment for this should
be explored.}
}
\value{
a \code{data.table} containing the empirical coverage for each quantile
in \code{predictions}.
}
\description{
This function computes empirical coverage for each quantile
in \code{predictions} and given observations \code{data} that should include a common
column of indexes for merging, which may be specified by \code{valid_time_index}.
}
