% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sr_progress_bar}
\alias{sr_progress_bar}
\title{A simple progress bar for package}
\usage{
sr_progress_bar(current, total, bar_length = 50)
}
\arguments{
\item{current}{numeric for the current iteration.}

\item{total}{total number of iterations.}

\item{bar_length}{length of the bart displayed in the console.}
}
\description{
A simple progress bar for package
}
\details{
Internal use.
}
